Reusable loan funded filter
* Cleaned up reused loan funded filter from components
* Added util file to replace mixin.
* Cleaned up components that were using the loan funded filter
